[refactor] rename commands to command
