+++ src/main/kotlin/command/Init.kt
@@ -1,11 +1,13 @@
-package com.an5on.commands
+package com.an5on.command
 
-import com.an5on.commands.GitUtils.buildCredentialsProvider
-import com.an5on.commands.GitUtils.parseRepoUrl
-import com.an5on.commands.GitUtils.remoteRepoPatterns
-import com.an5on.commands.GitUtils.sshSessionFactory
+import com.an5on.command.GitUtils.buildCredentialsProvider
+import com.an5on.command.GitUtils.parseRepoUrl
+import com.an5on.command.GitUtils.remoteRepoPatterns
+import com.an5on.command.GitUtils.sshSessionFactory
 import com.an5on.config.Configuration
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.Context
+import com.github.ajalt.clikt.core.installMordantMarkdown
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.optional
 import com.github.ajalt.clikt.parameters.options.flag
@@ -30,6 +32,10 @@ private val logger = KotlinLogging.logger {}
  * @author Anson Ng
  */
 class Init : CliktCommand() {
+    init {
+        installMordantMarkdown()
+    }
+
     val repo: String? by argument(
         help = "Clone from the specified remote Punkt repository"
     ).optional()
@@ -37,11 +43,33 @@ class Init : CliktCommand() {
         help = "Use SSH for cloning"
     ).flag()
     val branch: String? by option(
+        names = arrayOf("-b", "--branch"),
         help = "Set the branch of the remote Punkt repository to clone"
     )
     val depth: Int? by option(
+        names = arrayOf("-p", "--depth"),
         help = "Clone the remote Punkt repository shallowly with the specified depth"
     ).int()
+
+override fun help(context: Context) = """
+    Initialises a Punkt local repository at ${Configuration.localDirAbsPath}.
+            
+    If a remote Punkt repository URL is provided, it clones the repository to ${Configuration.localDirAbsPath}.
+    Otherwise, it initialises an empty Punkt local repository at ${Configuration.localDirAbsPath}.
+    
+    If the repository URL is not complete, Punkt will try to make guesses of it.
+    Supported formats for the remote Punkt repository URL:
+    ${remoteRepoPatterns.joinToString("\n") { "- ${it.pattern}" }}
+    
+    Examples:
+        punkt init
+        punkt init audrey
+        punkt init audrey/dotfiles
+        punkt init --ssh audrey/config
+        punkt init --branch main audrey/dotfiles
+        punkt init --depth 1 audrey/dotfiles
+        punkt init --ssh --branch main --depth 1 audrey/dotfiles
+    """.trimIndent()
 
     override fun run() {
         if (checkLocalExists()) {
